linera_ethereum

Module common

Common types and helpers shared across the crate, including error types and Ethereum event parsing.

Structs

EthereumEvent

The data type for an Ethereum event emitted by a smart contract

Enums

EthereumDataType

A single primitive data type. This is used for example for the entries of Ethereum events.

EthereumQueryError

An error that occurs when validating a JSON-RPC response.

EthereumServiceError

An error that occurs while accessing or parsing data from an Ethereum node.

Functions

event_name_from_expanded

Converts an entry named Event(type1 indexed,type2 indexed) into Event(type1,type2). event_name_expanded is needed for parsing the obtained log.

parse_log

Parses a log into an EthereumEvent using the expanded event signature event_name_expanded (e.g. MyEvent(type1 indexed,type2)).