Function linera_ethereum::test_utils::EventNumericsContract::deploy_builder

pub fn deploy_builder<T: Transport + Clone, P: Provider<T, N>, N: Network>(
    provider: P,
    initialSupply: U256,
) -> RawCallBuilder<T, P, N>

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

This is a simple wrapper around creating a RawCallBuilder with the data set to the bytecode concatenated with the constructor’s ABI-encoded arguments.