Module linera_ethereum::test_utils::SimpleTokenContract

Generated by the following Solidity interface…

interface SimpleTokenContract {
    event Initial(address, uint256 amount);
    event Transfer(address indexed from, address indexed to, uint256 amount);

    constructor(uint256 initialSupply);

    function balanceOf(address account) external view returns (uint256);
    function balances(address) external view returns (uint256);
    function totalSupply() external view returns (uint256);
    function transfer(address to, uint256 value) external returns (bool);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "initialSupply",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "balanceOf",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "balances",
    "inputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "totalSupply",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "transfer",
    "inputs": [
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "value",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "Initial",
    "inputs": [
      {
        "name": "",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Transfer",
    "inputs": [
      {
        "name": "from",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "to",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  }
]

Structs

Initial

Event with signature Initial(address,uint256) and selector 0x6a4da1829040b64159d6e13bc6bb2f42edf680a28387dc45b23b581e2942a2fe.

SimpleTokenContractInstance

A SimpleTokenContract instance.

Transfer

Event with signature Transfer(address,address,uint256) and selector 0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef.

balanceOfCall

Function with signature balanceOf(address) and selector 0x70a08231.

balanceOfReturn

Container type for the return parameters of the balanceOf(address) function.

balancesCall

Function with signature balances(address) and selector 0x27e235e3.

balancesReturn

Container type for the return parameters of the balances(address) function.

constructorCall

Constructor`.

totalSupplyCall

Function with signature totalSupply() and selector 0x18160ddd.

totalSupplyReturn

Container type for the return parameters of the totalSupply() function.

transferCall

Function with signature transfer(address,uint256) and selector 0xa9059cbb.

transferReturn

Container type for the return parameters of the transfer(address,uint256) function.

Enums

SimpleTokenContractCalls

Container for all the SimpleTokenContract function calls.

SimpleTokenContractEvents

Container for all the SimpleTokenContract events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain SimpleTokenContract contract instance.