Module linera_ethereum::test_utils::simple_contract

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

BalanceOfCall

Container type for all input parameters for the balanceOf function with signature balanceOf(address) and selector 0x70a08231

BalanceOfReturn

Container type for all return fields from the balanceOf function with signature balanceOf(address) and selector 0x70a08231

BalancesCall

Container type for all input parameters for the balances function with signature balances(address) and selector 0x27e235e3

BalancesReturn

Container type for all return fields from the balances function with signature balances(address) and selector 0x27e235e3

InitialFilter

SimpleContract

TotalSupplyCall

Container type for all input parameters for the totalSupply function with signature totalSupply() and selector 0x18160ddd

TotalSupplyReturn

Container type for all return fields from the totalSupply function with signature totalSupply() and selector 0x18160ddd

TransferCall

Container type for all input parameters for the transfer function with signature transfer(address,uint256) and selector 0xa9059cbb

TransferFilter

TransferReturn

Container type for all return fields from the transfer function with signature transfer(address,uint256) and selector 0xa9059cbb

TypesFilter

Enums

SimpleContractCalls

Container type for all of the contract’s call

SimpleContractEvents

Container type for all of the contract’s events

Statics

SIMPLECONTRACT_ABI

The parsed JSON ABI of the contract.

SIMPLECONTRACT_BYTECODE

The bytecode of the contract.

SIMPLECONTRACT_DEPLOYED_BYTECODE

The deployed bytecode of the contract.