Crate linera_ethereum

This module provides functionalities for accessing an Ethereum blockchain node. Enabling the ethereum allows to make the tests work. This requires installing the anvil from FOUNDRY and the SOLC compiler version 0.8.25

Modules

client

common

provider

test_utils

Helper types for tests and similar purposes.