linera_ethereum

Module common

Structs

EthereumEvent

The data type for an Ethereum event emitted by a smart contract

Enums

EthereumDataType

A single primitive data type. This is used for example for the entries of Ethereum events.

EthereumQueryError

EthereumServiceError

Functions

event_name_from_expanded

Converts an entry named Event(type1 indexed,type2 indexed) into Event(type1,type2). event_name_expanded is needed for parsing the obtained log.

parse_log