linera_ethereum::test_utils

Module EventNumericsContract

Generated by the following Solidity interface…

interface EventNumericsContract {
    event Types(address indexed from, address to, uint256 val0, uint64 val1, int64 val2, uint32 val3, int32 val4, uint16 val5, int16 val6, uint8 val7, int8 val8, bool val9);

    constructor(uint256 initialSupply);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "initialSupply",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "Types",
    "inputs": [
      {
        "name": "from",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "to",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "val0",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "val1",
        "type": "uint64",
        "indexed": false,
        "internalType": "uint64"
      },
      {
        "name": "val2",
        "type": "int64",
        "indexed": false,
        "internalType": "int64"
      },
      {
        "name": "val3",
        "type": "uint32",
        "indexed": false,
        "internalType": "uint32"
      },
      {
        "name": "val4",
        "type": "int32",
        "indexed": false,
        "internalType": "int32"
      },
      {
        "name": "val5",
        "type": "uint16",
        "indexed": false,
        "internalType": "uint16"
      },
      {
        "name": "val6",
        "type": "int16",
        "indexed": false,
        "internalType": "int16"
      },
      {
        "name": "val7",
        "type": "uint8",
        "indexed": false,
        "internalType": "uint8"
      },
      {
        "name": "val8",
        "type": "int8",
        "indexed": false,
        "internalType": "int8"
      },
      {
        "name": "val9",
        "type": "bool",
        "indexed": false,
        "internalType": "bool"
      }
    ],
    "anonymous": false
  }
]

Structs

EventNumericsContractInstance

A EventNumericsContract instance.

Types

Event with signature Types(address,address,uint256,uint64,int64,uint32,int32,uint16,int16,uint8,int8,bool) and selector 0x41567875ebd893ca4cdc8713e3b39234e8b097ba49316ece717044cb66137977.

constructorCall

Constructor`.

Enums

EventNumericsContractEvents

Container for all the EventNumericsContract events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain EventNumericsContract contract instance.